+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Sitemap controller.
+- Google Play Store link.
+- Video in project page.
 
 ## [0.0.8] - 2019-08-16
 ### Added